+++ exam/FinalTermOS/changeLog.md
@@ -3,3 +3,14 @@
 ## 2025 05 18 
 
 - Finish SleepingTA and DiningPhilosophers (PPT didn't make)
+
+## 2025 05 27
+
+- Student, StudentStatusDto, SleepingTASimulationStatus => SleepingTA
+- Chopstick, DiningPhilosophersSimulationStatus, Philosopher => DiningPhilosopher
+- SimulationStateEnums => For Q1 & Q2
+- BankersAlgorithmStatus, ProcessData, ResourceVector => BankersAlgorithm
+- Page, PageFrame, PageReplacementResult, VirtualMemorySimulationStatus => PageReplacement
+
+- Currently all questions are solved, but need further improvement and UI optimization.
+- Need to make slides to explain why and how the questions website are done.